v14

+++ include/dbs/NewProtocolReader.hpp
@@ -8,8 +8,8 @@
 
 bool readPulseBlockNewProtocol(const Config &cfg,
                                int beamskip,
-                               std::vector<std::complex<double>> &data1,
-                               std::vector<std::complex<double>> &data2,
+                               std::vector<std::complex<float>> &data1,
+                               std::vector<std::complex<float>> &data2,
                                std::vector<double> &utc,
                                double &theta_sq,
                                std::vector<std::vector<double>> &posRaw);
